pyproject.toml
[project]
name = "rejuganbot"
version = "1.0.0"
description = ""
authors = [
    {name = "Javier Gonel", email = "bolibic@gmail.com"},
]
dependencies = [
    "python-telegram-bot>=20.7",
    "pydantic-settings>=2.1.0",
    "aiosqlite>=0.19.0",
    "beautifulsoup4>=4.12.2",
    "aiohttp>=3.9.1",
    "fire>=0.5.0",
    "fastapi>=0.105.0",
    "jinja2>=3.1.2",
    "sqlalchemy[asyncio]>=2.0.23",
    "alembic>=1.13.1",
    "python-dateutil>=2.8.2",
    "tabulate>=0.9.0",
    "validators>=0.22.0",
    "uvicorn>=0.25.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[tool.pdm]

[tool.pdm.scripts]
telegram.call = "rejubot.telegrambot:run"
telegram.env_file = ".env"
web.cmd = "uvicorn rejubot.web:app --port 8080 --reload"
web.env_file = ".env"
alembic.cmd = "alembic"
alembic.env_file= ".env"
admin.call = "rejubot.admin:main"
admin.env_file = ".env"


[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.4.3",
    "ruff>=0.1.9",
]


[project.scripts]
rejubot = "rejubot.telegrambot:run"
rejubot-admin = "rejubot.admin:main"


